\name{SLC-package}
\alias{SLC-package}
\alias{SLC}
\docType{package}
\title{
Estimating the slope and level change
}
\description{
SLC is a package that estimates the slope and level change present in data after removing phase A trend. Represents graphically the original and the detrended data.
}
\details{
\tabular{ll}{
Package: \tab SLC\cr
Version: \tab 0.1\cr
Date: \tab 2010-14-01\cr
Depends: \tab >= 2.6.1\cr
License: \tab GPL version 2 or newer\cr
}

Index:
\preformatted{
slcestimates            Slope and level change estimates
}
}
\author{
Antonio Solanas <antonio.solanas@ub.edu>, Rumen Manolov <rrumenov13@ub.edu> & Patrick Onghena <patrick.onghena@ped.kuleuven.be>.

Maintainer: Rumen Manolov <rrumenov13@ub.edu>
}
\references{ Solanas, A. Manolov, R., & Onghena, P. (2009). Estimating slope and level change in single-case designs. Unpublished Manuscript. }
\keyword{package}
\seealso{
For more information see \code{\link[RcmdrPlugin.SLC]{RcmdrPlugin.SLC-package}}.
}
